Have plugin description account for the 2px font kern.
This constant is embedded in the Font.cpp, so we don't have direct
access to it here. It is used to force the rendering to start at the
initial baseline, but letters that dip below the baseline can be
rendered below the normal bounding box. We need to account for that here
so that we don't clip our p's and q's.