Unsplash sourcing batch 7
Sourced 10 unsplash images and stated whether they had the unsplash license or were public domain, and if they had the unsplash license, whether the terms allow selling unaltered copies. Found that another unsplash image was likely removed from unsplash and couldn't be sourced, and noted that in the copyright file.