Update news.txt
Added a small set of news for the CCOR region.
Bottom-most news, the one for Alvorada specifically, is incomplete. I've intended to finish it but can't seem to add more to it... assistance would be greatly appreciated.

Gagarin and Mordente-Bridi don't have unique news, this is not intentional rather it's a result of the writer's block that has halted progress on Alvorada's news. I would appreciate suggestions/contributions for their own news.