Allow missions to ping systems of importance
Pinged systems don't need to be visited in order for the mission to be completed.

Currently, pings display the same as stopovers or waypoints. (There's not enough space in the mission marker key to add another line for a new color.)

Pings are not removed when the system is visited. Pings can only be removed by the "unping" keyword in a GameAction.