Add an "escort payment" Mission attribute that determines the pay per escort per jump
By default, escorts still won't add any payment per jump.

If "escort payment" is present, the payload of the mission, which is normally cargo * 10 + passengers, gets the number of escorts * the escort payment added to it. Note that the payload is multiplied by the payment multiplier (typically 150) alongside the number of jumps to the destination to determine the final mission payment.

Note that we don't know how many escort ships are present until after all NPCs are instantiated, but the instantiation of NPCs uses the mission payload to determine the payments of actions within NPCs. As such, each NPC action will only take escort payment into consideration for the escorts in that NPC. For example, if you have two NPC blocks that are both escorts where one has 2 ships and the other has 3, then actions from the mission will add 5 * escort payment to the normal mission payload, but each NPC will only add 2 * escort payment and 3 * escort payment to the normal payload respectively.

This also fixes a bug where the `<npc>` text replacement wouldn't work in NPC action text, as actions were being instantiated before the NPC's ships were.