feat(balance): Adjusting mass, thrust, and turning across all ships (#34)
Co-authored-by: arkhne <36457239+arkhne@users.noreply.github.com>

Many thanks to Arkhne for doing the heavy lifting of all the calculations and working out variables.

These are neither final nor exact; but are general ballparks that are good enough to work with. I anticipate likely slowing down a lot of stuff, since we have easy access to a way to speed up the game; but no easy way to slow it down. It *may* be a bit too fast currently.