Reduce combat rating kestrel.txt
Simply reduces the combat rating required for the Kestrel mission to trigger from 8000 >> 6000.

As discussed in https://github.com/endless-sky/endless-sky/discussions/9980, I think the Kestrel mission comes too late. I don't agree with Saugia's comment on https://github.com/endless-sky/endless-sky/pull/8578#issuecomment-1484178674 that "It's not designed to be rewarding to the player in a progressive way based on their place in campaigns or combat in the game." Why not? It's a nice progression from a Falcon/Leviathan, a nice alternative to a Bactrian, and is still outclassed by the various alien ships the player will gain access to. 

Waiting until a combat rating of 8K basically guarantees the ship will be a fairly useless novelty, which I think is a waste.

Even if a new player has a combat rating of 6K then they still have to find the mission on Tarazad.