Feat(data): Implement Engine Slot system in Delta (#46)
This implements several things:
a) An engine slot system, whereby ships are limited to a certain number of thruster and steering components. This is to allow ships to actually be differentiated. As in, some ships might be able to do much better at steering, and others much better at accelerating, etc. Coalition ships and outfits, for instance, are modular and thus stack very well, which is now actually different than every other faction, since up until now, *everyone* 's ships and outfits stacked just as well, so wasn't much of a neat thing for them.
b) Hardpoint Info panel. I'm not set on the name, but in any case, this is a panel with the weapon loadout sprite and all the capacities and slots. This splits off from the Ship Info panel, which is now reserved for more operational things like performance stats and heat & energy balances.
c) The ship stat display system have, under the hood, been completely revamped. No longer is one single method pulling *all* the stats up for the Ship Info panel, but instead everything is being called up by about a dozen different methods. This means that the Ship Info and Hardpoint panel can be easily adjusted by coders. If someone wanted to add a fourth panel, or add a section of the stats somewhere else (maybe a panel hanging off the side of the Ship Info panel?), then it would be a matter of a few minutes to do so. Probably take longer to create the new panel graphics than to do the code.