feat(data): Scaling Sol (and potentially other interesting things) (#52)
This makes Sol properly-ish scaled at a scale of 1 ES unit = 100'000km. All the major planets are orbiting at a distance equal to their *average* orbital distance. I have not adjusted any of the moons, though. Someone taking a look at those would be appreciated, and implemented if it looks decent. 

This makes all the planets in Sol visitable, as well as a bunch of the moons. Many are inhabited, at least with small outposts, but many are uninhabited as well. Several Easter Eggs in there, too.

This also pushes the fence out to 64'000 in order to allow ships to actually reach Pluto.
This also sets the "universal ramscoop" to "false" because fuel is sufficiently trivialized as is, it doesn't need to be even more so.