Gentling the Zoom
This is both a safety and aesthetic factor.

Zooming farther out than 0.35 results in a diminishing return. Meaning the starfield and haze don't zoom out as fast, and prevents significant amounts of tiling even at extreme distances.

On my computer, this is the difference between 2541% GPU usage, and <50% GPU usage out at zoom values of 0.011