Added ability to select landable planets with the number keys.
Based on the same functionality in the EV series. Number keys 1-9 select the 1st to 9th landable planet or wormhole. The L command is unchanged and has to be used to actually order a landing.

To prevent this new functionality from interfering with players who may have set a number key as a command shortcut, it checks if a number key is in use and skips it. So eg if you have 1 assigned to some command, and you're in Sol, 2 now selects Earth, 3 Luna, and 4 Mars.