Remove unused HasCondition functionality
The keyword `has` for conditions is defined as "is not zero", which
means that actual presence of conditions is never visibile outside
the codebase. It also means that conditions which are zero don't need
to be written at saving, which we actually also don't do.

There were 3 locations within the codebase that used Has to check if
a condition was existing, but all those locations could easily be
implemented by Get, since they used a default value of 0 when the
condition was not present (which is exactly what Get also does).