Added 25 photos to Land
The files marked with the suffix -DAK are mine. The files marked with -Senn belong to Senn on discord and are submitted here with their permission.

These photos are all cropped to fit into the 720x360 resolution. I can provide them in the initial aspect ratios of you would like.
I can also edit out the airplane in clouds10, I just did not for the submission because it would take a while.