"never disabled" directly on ship definitions.
The "never disabled" keyword is not carried over, so it needs to be
directly on the ships in the person-section.

Also added "never disabled" to the regular Haulers as a workaround
for those haulers being capturable and then suddenly returning to
the author government. (An issue is filed to solve this properly.)