fix(system): Fix a rare segfault during wormhole link generation (#10333)
The segfault would occur when generating links from a legacy wormhole specification where the wormhole planet is undefined and data loaded after the original assignment to a stellar object removed it from a system. Because the planet was undefined, its "name" field would not map to it in the "planets" collection and so it was not correctly removed from the system.