feat(data): Expanding Alpha Centauri, and rearranging to be circumstellar rather than circumbinary (#73)
Alpha Centauri is an unusual system, as in it has three stars, and none of them are located at the center. It makes for an interesting counterpoint to Sol, which is very concentric. 

The space available in Alpha Centauri, and the fact that the planets are grouped together around the stars instead of directly around the center make for some interesting possibilities in terms of stellar system geometry and terrain.