feat(UI): Display Stars on the Galaxy Map (#93)
This adds a button to the map panel that displays the stars in each system. This allows players to see what stars are present and thus be able to make informed decisions about route finding when they are in a situation where solar and/or ramscoop resource production are important to them. It also makes the map much prettier. 

Note that this is not considered the end-state of this feature. The PR has some areas where it could be improved to satisfy more people, but these can be worked on in subsequent PR(s) to avoid causing problems with upstream.

Thank you to Azure3141 for all the work they put into making this! For transparency, note that I had no involvement in the creation of this code & data, it is Azure's work, with suggested code, suggestions, and input from a variety of other people.

Co-authored-by: Azure_ <42621251+Azure3141@users.noreply.github.com>
Co-authored-by: Rising Leaf <85687254+RisingLeaf@users.noreply.github.com>
Co-authored-by: TomGoodIdea <108272452+TomGoodIdea@users.noreply.github.com>
Co-authored-by: warp-core <warp-core@users.noreply.github.com>
Co-authored-by: tibetiroka <68112292+tibetiroka@users.noreply.github.com>