Add files via upload
Fixed issue whereby clicking on the 'buy' button when greyed out would only present one reason why a purchase could not be completed, even for situations where multiple reasons exist. This issue could be frustrating since the player could wrongly assume that resolving only the first issue would allow them to buy the item, only to find this was not the case. This issue existed because the code previously returned an error as soon as one was detected rather than accumulating them before returning at the end, this is now the case. In addition I have also clarified how many credits short the player is from making a purchase.