feat(data); Lampyrid Start (#116)
This adds a start where the player starts out with a Lampyrid Mk II spaceship and 480k loan, however 180k is lost in "fees". They still start in the same place, with a slightly altered James conversation. This is unlocked by having two years of in-game play on a single pilot. Intended to provide both an alternate start for testing, as well as to provide a different experience for veteran players.

Co-authored-by: Loymdayddaud <145969603+TheGiraffe3@users.noreply.github.com>