feat(data): Add a Hai person ship: Sanujab (#136)
Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>

This adds a small ship that is only found in Hai space. Unlike most person ships, this is not representative of a specific person, but rather serves to both expand the variety of ships in Hai space, and add a little more interest to the area.

This is the first of hopefully many new person ships that may or may not represent real people, with the dual intent of adding more variety to the game, as well as demonstrating more unique builds. (Only people creating person ships for themselves, and for completely fictional people will be accepted. No creating ships for other people unless said person independently confirms they agree to it. These persons ships should have at least a dozen lines of unique dialog suitable for the faction or area they are in, and should be vanilla ships where the only change is in how they are outfitted. Outfits are restricted to those available to the faction and area they are located in, with in general no spoiler or unique outfits.)