Feat(code): Add Intrasolar ship category & update flight check (#167)
Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>

Adds the ability to designate a ship (both via ship category and attribute that could be on an outfit) as intrasolar. Being so designated will suppress warnings about not having means of making jumps. 

Idea for a future PR: Add in specific warning and error icons to be put on ships to identify ships that lack means of making a jump. (it'd be a yellow warning for intrasolar ships, and a red error for non-intrasolar ships)