feat(code): Permanent changes to the player's ship via conversation actions. (#190)
Co-authored-by: vitalchip <65342577+vitalchip@users.noreply.github.com>

Missions can now use `attributes add hull 1300` style code to add attributes to the player's flagship, as well as `attributes set hull 1300` to set an attribute to something. Note this does not work properly with cost, and cannot change anything stored outside the `attribute` section. 

This code can be used both in an `on XXXX` section, as well as inside an action node of a conversation.