Feat(code): Add HUD display for Solar Power and Solar Wind (#164)
Co-authored-by: Loymdayddaud <145969603+TheGiraffe3@users.noreply.github.com>
Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>

This adds a new attribute that, when present on the player's flagship (typically via outfit, like the Astronomical Sensor added to the Advanced Deep outfitter), will allow the player to see the solar and wind absolute values of the system they are in, over by the radar. This allows the player to learn about these features even if they don't have space for solar panels or ramscoop right now; and also allows them to see what the potential is for this system; for instance, if they are far out and want to know if it is worth the risk of getting in closer to the system center.