Adding the Delta formation to most human fleets
This adds them to pretty much all merchant, Navy, and Security fleets.

Does *not* add them to pirate, CCOR, or independent fleets.

This just applies the "Delta (tailing)" formation to everything because it's one of the few that looks good and recognizeable even with really small 1-4 ship fleets, which most are.