feat(content): Monitor variant added to (still need to be renamed) Heimdall fleets/Added small fleet
Adds a new small fleet for the Heimdall, to keep the system populated yet not overwhelmingly so. Furthermore, adds a rare spawn of a Heimdall-refreshed Monitor.

This Monitor has been left MOSTLY alone, on account of just not being worth investing too many resources in. Still, they're reliable ships, so they see the odd use as a new AM boat.

Also, removes the (Warpless) tag from the Heimdall ships, as the warping fleet has now been completely removed. No need for it to stick around!