feat(content) Adjusted sales for anarchist lore
Adjusts the sales for the anarchist planets, making them better fit the lore of the two factions.

The True Anarchists, being the unorganized and scattered group, focus more on mining and non-combat outfits. They also sell "civilian" ships, like haulers and behemoths.

Meanwhile the Northern Anarchists, with more "organization" and concentrated colonies, plus having a lot of old factories to work with, manufacturer advanced ships and weapons without much of a focus on the more primitive and non-combat sales.