Further tweaked randomness and payments
Further tweaks the randomness to hopefully flood the job board a little less. Still possible! Just, not all the time now.

Also, adjusts prices to be a bit fairer, but leaves the luxury transports untouched. The goal here is to highlight a disparity between the average joe and the wealthy elite. Turns out, even in a lawless anarchy, pirate society isn't too different than the governed world in a social class respect...