Change how RouteCacheKeys are generated
The previous method could cause ships that should generate two different paths to use the same one, as it lacked information about the ship's personality and only used the ship's max jump range and whether it was capable of jumping.

I believe this now covers all possible jump-relevant information in the cache key.